Split meilisearch into user and document search classes
