+++ src/main/java/de/skillkiller/documentdbackend/service/UserDetailsService.java
@@ -2,7 +2,7 @@ package de.skillkiller.documentdbackend.service;
 
 import de.skillkiller.documentdbackend.entity.User;
 import de.skillkiller.documentdbackend.entity.UserDetailsHolder;
-import de.skillkiller.documentdbackend.search.MeiliSearch;
+import de.skillkiller.documentdbackend.search.UserSearch;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
@@ -13,15 +13,15 @@ import java.util.Optional;
 @Service
 public class UserDetailsService implements org.springframework.security.core.userdetails.UserDetailsService {
 
-    private final MeiliSearch meiliSearch;
+    private final UserSearch meiliSearch;
 
-    public UserDetailsService(MeiliSearch meiliSearch) {
+    public UserDetailsService(UserSearch meiliSearch) {
         this.meiliSearch = meiliSearch;
     }
 
     @Override
     public UserDetails loadUserByUsername(String s) throws UsernameNotFoundException {
-        Optional<User> optionalUser = meiliSearch.searchUserByUsername(s);
+        Optional<User> optionalUser = meiliSearch.getUserByUsername(s);
 
         if (optionalUser.isPresent()) {
             return new UserDetailsHolder(optionalUser.get(), Collections.emptyList());
